refactor: Move review repository to repository directory
Move review repository to repository directory
